scripts/requirements.txt
hsqldb>=2.5.0
jpype1>=1.4.1
